CasinoSceneChanges
Created Camera Bounds in the X and Y axis of the casino scene
created a layer called BlockRaycast that, when tapped, will not make the player start walking towards it
player movement tap will ignore objects in all layers except for BlockRaycast and only tap on things in the clickableObjects layer
Worked on UI Scaling in Casino Scene, but definitely needs work still.
Changed the requirement for the PromptController's scene change function. Instead of taking an int it now takes the string of the scene's name. (Thinking about this now, I want to make it a drop down menu instead)